pyproject.toml
[project]
name = "llmwikidoc"
version = "0.1.0"
description = "LLM-powered wiki that auto-updates from git commits (Karpathy LLM Wiki pattern)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.49",
    "google-genai>=1.74.0",
    "networkx>=3.6.1",
    "rank-bm25>=0.2.2",
    "rich>=15.0.0",
    "tomli>=2.4.1",
    "tomli-w>=1.2.0",
    "typer>=0.25.0",
]

[project.scripts]
llmwikidoc = "llmwikidoc.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmwikidoc"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
